Add reusable scene model symbol capture service
