+++ gui/tools/fixture_symbol_generation_tool.cpp
@@ -1,7 +1,6 @@
 #include "tools/fixture_symbol_generation_tool.h"
 
 #include <algorithm>
-#include <array>
 #include <map>
 #include <set>
 #include <string>
@@ -14,39 +13,13 @@
 #include "dialogs/GenerateFixtureSymbolsDialog.h"
 #include "guiconfigservices.h"
 #include "mainwindow.h"
-#include "sceneobject.h"
-#include "support.h"
-#include "truss.h"
 #include "opaque_pass_utils.h"
-#include "symbols/Symbol2DImageBuilder.h"
+#include "tools/scene_model_symbol_capture_service.h"
 #include "viewer2doffscreenrenderer.h"
-#include "viewer2dpanel.h"
 #include "windows/SymbolPreviewWindow.h"
 
 namespace tools {
 namespace {
-
-class ScopedFloatConfigOverride {
-public:
-  ScopedFloatConfigOverride(ConfigManager &cfg,
-                            std::initializer_list<std::pair<const char *, float>> values)
-      : cfg_(cfg) {
-    for (const auto &entry : values) {
-      const std::string key(entry.first);
-      previous_[key] = cfg_.GetFloat(key);
-      cfg_.SetFloat(key, entry.second);
-    }
-  }
-
-  ~ScopedFloatConfigOverride() {
-    for (const auto &[key, value] : previous_)
-      cfg_.SetFloat(key, value);
-  }
-
-private:
-  ConfigManager &cfg_;
-  std::unordered_map<std::string, float> previous_;
-};
 
 bool FixtureMatchesModelKeys(const Fixture &fixture,
                              const std::vector<std::string> &modelKeys) {
@@ -75,75 +48,6 @@ std::string FindSingleFixtureUuidForModelKeys(
   }
   return selectedUuid;
 }
-
-class ScopedFixtureColorOverride {
-public:
-  ScopedFixtureColorOverride(ConfigManager &cfg, const std::string &fixtureUuid,
-                             const std::string &forcedHex)
-      : cfg_(cfg) {
-    if (fixtureUuid.empty())
-      return;
-    auto &fixtures = cfg_.GetScene().fixtures;
-    auto it = fixtures.find(fixtureUuid);
-    if (it == fixtures.end())
-      return;
-    previous_.emplace_back(it->first, it->second.color);
-    it->second.color = forcedHex;
-  }
-
-  ~ScopedFixtureColorOverride() {
-    auto &fixtures = cfg_.GetScene().fixtures;
-    for (const auto &[uuid, color] : previous_) {
-      auto it = fixtures.find(uuid);
-      if (it != fixtures.end())
-        it->second.color = color;
-    }
-  }
-
-private:
-  ConfigManager &cfg_;
-  std::vector<std::pair<std::string, std::string>> previous_;
-};
-
-class ScopedSymbolCaptureSceneOverride {
-public:
-  ScopedSymbolCaptureSceneOverride(ConfigManager &cfg,
-                                   const std::string &fixtureUuid)
-      : cfg_(cfg) {
-    auto &scene = cfg_.GetScene();
-    originalFixtures_ = scene.fixtures;
-    originalTrusses_ = scene.trusses;
-    originalSceneObjects_ = scene.sceneObjects;
-    originalSupports_ = scene.supports;
-
-    std::unordered_map<std::string, Fixture> filteredFixtures;
-    if (!fixtureUuid.empty()) {
-      auto fixtureIt = scene.fixtures.find(fixtureUuid);
-      if (fixtureIt != scene.fixtures.end())
-        filteredFixtures.emplace(fixtureIt->first, fixtureIt->second);
-    }
-
-    scene.fixtures = std::move(filteredFixtures);
-    scene.trusses.clear();
-    scene.sceneObjects.clear();
-    scene.supports.clear();
-  }
-
-  ~ScopedSymbolCaptureSceneOverride() {
-    auto &scene = cfg_.GetScene();
-    scene.fixtures = std::move(originalFixtures_);
-    scene.trusses = std::move(originalTrusses_);
-    scene.sceneObjects = std::move(originalSceneObjects_);
-    scene.supports = std::move(originalSupports_);
-  }
-
-private:
-  ConfigManager &cfg_;
-  std::unordered_map<std::string, Fixture> originalFixtures_;
-  std::unordered_map<std::string, Truss> originalTrusses_;
-  std::unordered_map<std::string, SceneObject> originalSceneObjects_;
-  std::unordered_map<std::string, Support> originalSupports_;
-};
 
 std::vector<FixtureSymbolTypeOption> BuildFixtureOptions() {
   std::vector<FixtureSymbolTypeOption> options;
@@ -176,26 +80,6 @@ std::vector<FixtureSymbolTypeOption> BuildFixtureOptions() {
   return options;
 }
 
-void MirrorImageHorizontally(symbols::RenderedSymbolImage &render) {
-  if (render.width <= 0 || render.height <= 0)
-    return;
-  for (int y = 0; y < render.height; ++y) {
-    for (int x = 0; x < render.width / 2; ++x) {
-      const int opposite = render.width - 1 - x;
-      const size_t left =
-          (static_cast<size_t>(y) * static_cast<size_t>(render.width) +
-           static_cast<size_t>(x)) *
-          4;
-      const size_t right =
-          (static_cast<size_t>(y) * static_cast<size_t>(render.width) +
-           static_cast<size_t>(opposite)) *
-          4;
-      for (size_t c = 0; c < 4; ++c)
-        std::swap(render.rgba[left + c], render.rgba[right + c]);
-    }
-  }
-}
-
 } // namespace
 
 void RunFixtureSymbolGeneration(MainWindow &window) {
@@ -221,13 +105,6 @@ void RunFixtureSymbolGeneration(MainWindow &window) {
                  "Generate Fixture Symbols", wxOK | wxICON_ERROR, &window);
     return;
   }
-  Viewer2DPanel *capturePanel = offscreenRenderer->GetPanel();
-  if (!capturePanel) {
-    wxMessageBox("Could not create 2D capture panel instance.",
-                 "Generate Fixture Symbols", wxOK | wxICON_ERROR, &window);
-    return;
-  }
-
   ConfigManager &cfg = GetDefaultGuiConfigServices().LegacyConfigManager();
   const std::string selectedFixtureUuid =
       FindSingleFixtureUuidForModelKeys(cfg.GetScene().fixtures,
@@ -239,91 +116,20 @@ void RunFixtureSymbolGeneration(MainWindow &window) {
   }
 
   const std::string forcedFixtureColor = "#3FA9F5";
-  ScopedSymbolCaptureSceneOverride isolatedSceneOverride(cfg, selectedFixtureUuid);
-  ScopedFixtureColorOverride selectedFixtureColorOverride(cfg, selectedFixtureUuid,
-                                                          forcedFixtureColor);
-  ScopedFloatConfigOverride displayOverride(
-      cfg,
-      {
-          {"grid_show", 0.0f},
-          {"view2d_dark_mode", 0.0f},
-          {"view2d_render_mode",
-           static_cast<float>(Viewer2DRenderMode::ByFixtureType)},
-          {"label_show_name_top", 0.0f},
-          {"label_show_name_front", 0.0f},
-          {"label_show_name_side", 0.0f},
-          {"label_show_id_top", 0.0f},
-          {"label_show_id_front", 0.0f},
-          {"label_show_id_side", 0.0f},
-          {"label_show_dmx_top", 0.0f},
-          {"label_show_dmx_front", 0.0f},
-          {"label_show_dmx_side", 0.0f},
-      });
-
-  const Viewer2DView previousView = capturePanel->GetView();
-
-  offscreenRenderer->SetViewportSize(wxSize(1200, 1200));
-  offscreenRenderer->PrepareForCapture();
-  capturePanel->SetRenderMode(Viewer2DRenderMode::ByFixtureType);
-  capturePanel->UpdateScene(true);
-  {
-    std::vector<unsigned char> warmupPixels;
-    int warmupWidth = 0;
-    int warmupHeight = 0;
-    (void)capturePanel->RenderToRGBA(warmupPixels, warmupWidth, warmupHeight);
-  }
-
-  struct CaptureRequest {
-    Viewer2DView view = Viewer2DView::Top;
-    float topFixturesInverted = 0.0f;
-    bool mirrorSideHorizontally = false;
-    symbols::SymbolView symbolView = symbols::SymbolView::Top;
-  };
-  const std::array<CaptureRequest, 4> requests = {
-      CaptureRequest{Viewer2DView::Front, 0.0f, false, symbols::SymbolView::Front},
-      CaptureRequest{Viewer2DView::Top, 0.0f, false, symbols::SymbolView::Top},
-      CaptureRequest{Viewer2DView::Side, 0.0f, true, symbols::SymbolView::Left},
-      CaptureRequest{Viewer2DView::Top, 1.0f, false, symbols::SymbolView::Bottom},
-  };
-
-  std::vector<symbols::RenderedSymbolImage> renders;
-  renders.reserve(requests.size());
-  bool allOk = true;
-  for (const auto &request : requests) {
-    ScopedFloatConfigOverride topViewOverride(
-        cfg, {{"view2d_top_fixtures_inverted", request.topFixturesInverted}});
-
-    capturePanel->SetView(request.view);
-    capturePanel->UpdateScene(true);
-    capturePanel->FitViewToScene();
-
-    symbols::RenderedSymbolImage render;
-    render.view = request.symbolView;
-    bool ok = capturePanel->RenderToRGBA(render.rgba, render.width, render.height);
-    if (!ok || render.width <= 0 || render.height <= 0) {
-      allOk = false;
-      break;
-    }
-    if (request.mirrorSideHorizontally)
-      MirrorImageHorizontally(render);
-    renders.push_back(std::move(render));
-  }
-  capturePanel->SetView(previousView);
-
-  if (!allOk) {
-    wxMessageBox("Could not capture all orthographic source images from the 2D viewer.",
-                 "Generate Fixture Symbols", wxOK | wxICON_ERROR, &window);
+  SceneModelSymbolCaptureOptions captureOptions;
+  captureOptions.forcedFixtureColor = forcedFixtureColor;
+  auto capture = CaptureSceneModelOrthographicSymbols(
+      *offscreenRenderer, cfg,
+      SceneModelSymbolTarget{SceneModelKind::Fixture, selectedFixtureUuid},
+      captureOptions);
+  if (!capture.ok) {
+    wxMessageBox(capture.error, "Generate Fixture Symbols", wxOK | wxICON_ERROR,
+                 &window);
     return;
   }
 
-  auto symbols = symbols::Symbol2DImageBuilder::BuildFromRenderedImages(renders);
-  if (symbols.size() != requests.size()) {
-    wxMessageBox("Could not generate all fixture symbols from captured views.",
-                 "Generate Fixture Symbols", wxOK | wxICON_ERROR, &window);
-    return;
-  }
-
-  SymbolPreviewWindow *preview = new SymbolPreviewWindow(&window, std::move(symbols));
+  SymbolPreviewWindow *preview =
+      new SymbolPreviewWindow(&window, std::move(capture.symbols));
   preview->Show();
 }
